requirements.txt
yfinance~=0.2.11
python-dotenv~=0.21.1
aiogram~=2.25.1
python-coinmarketcap~=0.5
